requirements.txt
discord.py
Flask
mistralai